TINY-10022: move createNewBlock to NewLineUtils
